redesign onboard flow
We now go from the initial setup to the advanced setup, but with a BIG
button to take the user to the tar1090 page. This will also add
hopefully enough delay to ensur that the tar1090 server is up and ready
by the time the user tries to access the map.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>